Implemented the code for refills and fixed a bug
- bug: TaxiExample callForTaxi method had a loop which increased the
radius if the initial call was unsuccessful; but this was happening
within a single "tick" and if there were no free taxis on the map (no
matter how big the radius) the program would go into an infinite loop.
Fix: callForTaxi does only one try (the radius is given). ticklistener
tick method increases the radius size on the next tick and keeps a list
of customers that haven't been assigned to a taxi yet
(bufferedCustomers)